+++ ndnSimV24/ns-3/src/ndnSIM/helper/ndn-fib-helper.cpp
@@ -131,9 +131,9 @@ FibHelper::AddRoute(Ptr<Node> node, const Name& prefix, shared_ptr<Face> face, i
 //Dome
 void FibHelper::AddRoute(Ptr<Node> node, const Name& prefix,
 		shared_ptr<Face> face, int32_t metric, std::string macAddress,
-		double positionX, double positionY, double positionZ,
-		double futurePositionX, double futurePositionY,
-		double timeAtFuturePosition, double isFuturePositionSet) {
+		int positionX, int positionY, int positionZ,
+		int futurePositionX, int futurePositionY,
+		int timeAtFuturePosition, int isFuturePositionSet) {
 	NS_LOG_LOGIC(
 			"[" << node->GetId() << "]$ route add " << prefix << " via " << face->getLocalUri() << " metric " << metric);
 
@@ -191,9 +191,9 @@ FibHelper::AddRoute(Ptr<Node> node, const Name& prefix, uint32_t faceId, int32_t
 }
 //Dome
 void FibHelper::AddRoute(Ptr<Node> node, const Name& prefix, uint32_t faceId,
-		int32_t metric, std::string macAddress, double positionX,
-		double positionY, double positionZ, double futurePositionX,
-		double futurePositionY, double timeAtFuturePosition, double isFuturePositionSet) {
+		int32_t metric, std::string macAddress, int positionX,
+		int positionY, int positionZ, int futurePositionX,
+		int futurePositionY, int timeAtFuturePosition, int isFuturePositionSet) {
 	Ptr<L3Protocol> ndn = node->GetObject<L3Protocol>();
 	NS_ASSERT_MSG(ndn != 0, "Ndn stack should be installed on the node");
 
@@ -241,9 +241,9 @@ FibHelper::AddRoute(const std::string& nodeName, const Name& prefix, uint32_t fa
 //Dome
 void FibHelper::AddRoute(const std::string& nodeName, const Name& prefix,
 		uint32_t faceId, int32_t metric, std::string macAddress,
-		double positionX, double positionY, double positionZ,
-		double futurePositionX, double futurePositionY,
-		double timeAtFuturePosition, double isFuturePositionSet) {
+		int positionX, int positionY, int positionZ,
+		int futurePositionX, int futurePositionY,
+		int timeAtFuturePosition, int isFuturePositionSet) {
 	Ptr<Node> node = Names::Find<Node>(nodeName);
 	NS_ASSERT_MSG(node != 0, "Node [" << nodeName << "] does not exist");
 
@@ -323,9 +323,9 @@ FibHelper::AddRoute(Ptr<Node> node, const Name& prefix, Ptr<Node> otherNode, int
 //Dome
 void FibHelper::AddRoute(Ptr<Node> node, const Name& prefix,
 		Ptr<Node> otherNode, int32_t metric, std::string macAddress,
-		double positionX, double positionY, double positionZ,
-		double futurePositionX, double futurePositionY,
-		double timeAtFuturePosition, double isFuturePositionSet) {
+		int positionX, int positionY, int positionZ,
+		int futurePositionX, int futurePositionY,
+		int timeAtFuturePosition, int isFuturePositionSet) {
 	for (uint32_t deviceId = 0; deviceId < node->GetNDevices(); deviceId++) {
 		Ptr<PointToPointNetDevice> netDevice =
 				DynamicCast<PointToPointNetDevice>(node->GetDevice(deviceId));
@@ -387,9 +387,9 @@ FibHelper::AddRoute(const std::string& nodeName, const Name& prefix,
 //Dome
 void FibHelper::AddRoute(const std::string& nodeName, const Name& prefix,
 		const std::string& otherNodeName, int32_t metric,
-		std::string macAddress, double positionX, double positionY,
-		double positionZ, double futurePositionX, double futurePositionY,
-		double timeAtFuturePosition, double isFuturePositionSet) {
+		std::string macAddress, int positionX, int positionY,
+		int positionZ, int futurePositionX, int futurePositionY,
+		int timeAtFuturePosition, int isFuturePositionSet) {
 	Ptr<Node> node = Names::Find<Node>(nodeName);
 	NS_ASSERT_MSG(node != 0, "Node [" << nodeName << "] does not exist");
 
@@ -433,9 +433,9 @@ FibHelper::RemoveRoute(Ptr<Node> node, const Name& prefix, shared_ptr<Face> face
 }
 //Dome
 void FibHelper::RemoveRoute(Ptr<Node> node, const Name& prefix,
-		shared_ptr<Face> face, std::string macAddress, double positionX,
-		double positionY, double positionZ, double futurePositionX,
-		double futurePositionY, double timeAtFuturePosition, double isFuturePositionSet) {
+		shared_ptr<Face> face, std::string macAddress, int positionX,
+		int positionY, int positionZ, int futurePositionX,
+		int futurePositionY, int timeAtFuturePosition, int isFuturePositionSet) {
 	// Get L3Protocol object
 	Ptr<L3Protocol> L3protocol = node->GetObject<L3Protocol>();
 	// Get the forwarder instance
@@ -483,9 +483,9 @@ FibHelper::RemoveRoute(Ptr<Node> node, const Name& prefix, uint32_t faceId, std:
 }
 //Dome
 void FibHelper::RemoveRoute(Ptr<Node> node, const Name& prefix, uint32_t faceId,
-		std::string macAddress, double positionX, double positionY,
-		double positionZ, double futurePositionX, double futurePositionY,
-		double timeAtFuturePosition, double isFuturePositionSet) {
+		std::string macAddress, int positionX, int positionY,
+		int positionZ, int futurePositionX, int futurePositionY,
+		int timeAtFuturePosition, int isFuturePositionSet) {
 	Ptr<L3Protocol> ndn = node->GetObject<L3Protocol>();
 	NS_ASSERT_MSG(ndn != 0, "Ndn stack should be installed on the node");
 
@@ -527,9 +527,9 @@ FibHelper::RemoveRoute(const std::string& nodeName, const Name& prefix, uint32_t
 }
 //Dome
 void FibHelper::RemoveRoute(const std::string& nodeName, const Name& prefix,
-		uint32_t faceId, std::string macAddress, double positionX,
-		double positionY, double positionZ, double futurePositionX,
-		double futurePositionY, double timeAtFuturePosition, double isFuturePositionSet) {
+		uint32_t faceId, std::string macAddress, int positionX,
+		int positionY, int positionZ, int futurePositionX,
+		int futurePositionY, int timeAtFuturePosition, int isFuturePositionSet) {
 	Ptr<Node> node = Names::Find<Node>(nodeName);
 	Ptr<L3Protocol> ndn = node->GetObject<L3Protocol>();
 	NS_ASSERT_MSG(ndn != 0, "Ndn stack should be installed on the node");
@@ -606,9 +606,9 @@ FibHelper::RemoveRoute(Ptr<Node> node, const Name& prefix, Ptr<Node> otherNode, 
 //Dome
 
 void FibHelper::RemoveRoute(Ptr<Node> node, const Name& prefix,
-		Ptr<Node> otherNode, std::string macAddress, double positionX,
-		double positionY, double positionZ, double futurePositionX,
-		double futurePositionY, double timeAtFuturePosition, double isFuturePositionSet) {
+		Ptr<Node> otherNode, std::string macAddress, int positionX,
+		int positionY, int positionZ, int futurePositionX,
+		int futurePositionY, int timeAtFuturePosition, int isFuturePositionSet) {
 	for (uint32_t deviceId = 0; deviceId < node->GetNDevices(); deviceId++) {
 		Ptr<PointToPointNetDevice> netDevice =
 				DynamicCast<PointToPointNetDevice>(node->GetDevice(deviceId));
@@ -661,9 +661,9 @@ FibHelper::RemoveRoute(const std::string& nodeName, const Name& prefix,
 //Dome
 void FibHelper::RemoveRoute(const std::string& nodeName, const Name& prefix,
 		const std::string& otherNodeName, std::string macAddress,
-		double positionX, double positionY, double positionZ,
-		double futurePositionX, double futurePositionY,
-		double timeAtFuturePosition, double isFuturePositionSet) {
+		int positionX, int positionY, int positionZ,
+		int futurePositionX, int futurePositionY,
+		int timeAtFuturePosition, int isFuturePositionSet) {
 	Ptr<Node> node = Names::Find<Node>(nodeName);
 	Ptr<Node> otherNode = Names::Find<Node>(otherNodeName);
 	RemoveRoute(node, prefix, otherNode, macAddress, positionX, positionY,
