+++ ndnSimV24/ns-3/src/ndnSIM/helper/ndn-fib-helper.hpp
@@ -58,8 +58,8 @@ public:
   //Dome
   static void
   	AddRoute(Ptr<Node> node, const Name& prefix, shared_ptr<Face> face,
-  			int32_t metric, std::string macAddress, double positionX,
-  			double positionY , double positionZ, double futurePositionX, double futurePositionY, double timeAtFuturePosition, double isFuturePositionSet);
+  			int32_t metric, std::string macAddress, int positionX,
+  			int positionY , int positionZ, int futurePositionX, int futurePositionY, int timeAtFuturePosition, int isFuturePositionSet);
 
   /**
    * \brief Add forwarding entry to FIB
@@ -76,8 +76,8 @@ public:
   //Dome
   static void
   	AddRoute(const std::string& nodeName, const Name& prefix, uint32_t faceId,
-  			int32_t metric, std::string macAddress, double positionX,
-  			double positionY , double positionZ, double futurePositionX, double futurePositionY, double timeAtFuturePosition, double isFuturePositionSet);
+  			int32_t metric, std::string macAddress, int positionX,
+  			int positionY , int positionZ, int futurePositionX, int futurePositionY, int timeAtFuturePosition, int isFuturePositionSet);
 
   /**
    * \brief Add forwarding entry to FIB
@@ -94,9 +94,9 @@ public:
   //Dome
 	static void
 	AddRoute(Ptr<Node> node, const Name& prefix, uint32_t faceId,
-			int32_t metric, std::string macAddress, double positionX,
-			double positionY, double positionZ, double futurePositionX,
-			double futurePositionY, double timeAtFuturePosition, double isFuturePositionSet);
+			int32_t metric, std::string macAddress, int positionX,
+			int positionY, int positionZ, int futurePositionX,
+			int futurePositionY, int timeAtFuturePosition, int isFuturePositionSet);
 
   /**
    * @brief Add forwarding entry to FIB (work only with point-to-point links)
@@ -113,9 +113,9 @@ public:
   //Dome
   static void
   	AddRoute(Ptr<Node> node, const Name& prefix, Ptr<Node> otherNode,
-  			int32_t metric, std::string macAddress, double positionX,
-  			double positionY, double positionZ, double futurePositionX,
-  			double futurePositionY, double timeAtFuturePosition, double isFuturePositionSet);
+  			int32_t metric, std::string macAddress, int positionX,
+  			int positionY, int positionZ, int futurePositionX,
+  			int futurePositionY, int timeAtFuturePosition, int isFuturePositionSet);
 
   /**
    * @brief Add forwarding entry to FIB (work only with point-to-point links)
@@ -135,9 +135,9 @@ public:
   static void
   	AddRoute(const std::string& nodeName, const Name& prefix,
   			const std::string& otherNodeName, int32_t metric,
-  			std::string macAddress, double positionX, double positionY,
-  			double positionZ, double futurePositionX, double futurePositionY,
-  			double timeAtFuturePosition, double isFuturePositionSet);
+  			std::string macAddress, int positionX, int positionY,
+  			int positionZ, int futurePositionX, int futurePositionY,
+  			int timeAtFuturePosition, int isFuturePositionSet);
 
 
   /**
@@ -154,9 +154,9 @@ public:
   //Dome
 	static void
 	RemoveRoute(Ptr<Node> node, const Name& prefix, shared_ptr<Face> face,
-			std::string macAddress, double positionX, double positionY,
-			double positionZ, double futurePositionX, double futurePositionY,
-			double timeAtFuturePosition, double isFuturePositionSet);
+			std::string macAddress, int positionX, int positionY,
+			int positionZ, int futurePositionX, int futurePositionY,
+			int timeAtFuturePosition, int isFuturePositionSet);
 
   /**
    * \brief remove forwarding entry in FIB
@@ -172,9 +172,9 @@ public:
   //Dome
 	static void
 	RemoveRoute(Ptr<Node> node, const Name& prefix, uint32_t faceId,
-			std::string macAddress, double positionX, double positionY,
-			double positionZ, double futurePositionX, double futurePositionY,
-			double timeAtFuturePosition, double isFuturePositionSet);
+			std::string macAddress, int positionX, int positionY,
+			int positionZ, int futurePositionX, int futurePositionY,
+			int timeAtFuturePosition, int isFuturePositionSet);
 
   /**
    * \brief remove forwarding entry in FIB
@@ -190,9 +190,9 @@ public:
   //Dome
 	static void
 	RemoveRoute(const std::string& nodeName, const Name& prefix,
-			uint32_t faceId, std::string macAddress, double positionX,
-			double positionY, double positionZ, double futurePositionX,
-			double futurePositionY, double timeAtFuturePosition, double isFuturePositionSet);
+			uint32_t faceId, std::string macAddress, int positionX,
+			int positionY, int positionZ, int futurePositionX,
+			int futurePositionY, int timeAtFuturePosition, int isFuturePositionSet);
 
   /**
    * @brief remove forwarding entry in FIB (work only with point-to-point links)
@@ -208,9 +208,9 @@ public:
   //Dome
   static void
   	RemoveRoute(Ptr<Node> node, const Name& prefix, Ptr<Node> otherNode,
-  			std::string macAddress, double positionX, double positionY,
-  			double positionZ, double futurePositionX, double futurePositionY,
-  			double timeAtFuturePosition, double isFuturePositionSet);
+  			std::string macAddress, int positionX, int positionY,
+  			int positionZ, int futurePositionX, int futurePositionY,
+  			int timeAtFuturePosition, int isFuturePositionSet);
 
   /**
    * @brief remove forwarding entry in FIB (work only with point-to-point links)
@@ -227,9 +227,9 @@ public:
   static void
   	RemoveRoute(const std::string& nodeName, const Name& prefix,
   			const std::string& otherNodeName, std::string macAddress,
-  			double positionX, double positionY, double positionZ,
-  			double futurePositionX, double futurePositionY,
-  			double timeAtFuturePosition, double isFuturePositionSet);
+  			int positionX, int positionY, int positionZ,
+  			int futurePositionX, int futurePositionY,
+  			int timeAtFuturePosition, int isFuturePositionSet);
 
 private:
   static void
